Test de NotebookDAO. En el fichero de inicio (EverpobreApp) se inicializa la bbdd Se cambia el id de nuestro Notebook de un int a un long Se crea la clase DBConstant para la crear las ctes de las tablas que se usen en la app Se crea el fichero DBHelper para la creación, actualización de la bbdd Se crea el fichero NotebookDAO que sirve de puente entre la clase Notebook y la tabla de bbdd Notebook. Aquí se encuentran las tareas de query, inserta, update y delete

+++ Everpobre/app/src/main/java/fjc/com/everpobre/model/Notebook.java
@@ -5,10 +5,10 @@ package fjc.com.everpobre.model;
  */
 public class Notebook {
     public static final String DEFAULT_NAME = "John Doe";
-    private int id;
+    private long id;
     private String name;
 
-    public Notebook(int id, String name) {
+    public Notebook(long id, String name) {
 
         if (name == null || name.isEmpty()){
 
@@ -27,11 +27,11 @@ public class Notebook {
         //Desactivo el constructor por defecto
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
